+++ src/constants.js
@@ -61,11 +61,13 @@ export const DEFAULT_COMPILER_ID = NO_PLUGIN_TYPE;
 
 export const DEFAULT_TESTER_ID = NO_PLUGIN_TYPE;
 
+export const DEFAULT_PACKAGE_MANAGER = 'npm';
+
 export const DEFAULT_EXTENSIONS = {
   'ext-docs-parser': {
     config: {},
     options: {
-      default: true
+      core: true
     }
   }
 };